Doesn't work
Crashes with GC overhead error during setup on my machine.
If it did run, it would get *maybe* one article. Not quite sure what I'm
supposed to be doing on line 89. I'm thinking I'll need to manually
WikipediaPages from the <(/)page> tags, THEN use getTitle and
getContent. It shouldn't be hard, but first I need to *get to* to
mapper.