Working...ish
pom updated to make a fat jar with dependency jars included
people.txt included as a resource because I couldn't get it to work as a
third argument
GetArticlesMapred works... on trivially small datasets. On larger wiki
dumps, it throws a "bytes consumed" error, which, as far as I can tell,
super shouldn't be happening. It's only supposed to show when working
with gzipped files (we're not), and from the comments it seems like
debug code that should have been removed. I think I might be the first
person to ever actually see this error thrown.

I'm probably just using the wrong version of a library or something.
Maybe it'll work better on someone else's machine?